refactor(search): split compiler and duplicate helpers
Move query compilation and OCR duplicate suppression out of the main search pipeline so search execution can be reviewed independently from matcher setup and de-duplication rules.
