+++ src/core/search.ts
@@ -1,116 +1,13 @@
 import type { FormField, OcrWord, PageAnnotation, PageOcr, SearchMatch, TextSpan } from '../types/index.js';
 import { CJK_TIGHT_GAP_RATIO, isCjkLeading } from './cjkJoin.js';
+import { type CompiledSearch, nfkc } from './search/compiler.js';
+import { duplicateKey, hasPreciseDuplicateAtBox, suppressDuplicateOcrMatches } from './search/duplicates.js';
+
+export { type CompiledSearch, compileSearch } from './search/compiler.js';
+export { suppressDuplicateOcrMatches } from './search/duplicates.js';
+
 import { isLikelyCjkDisplaySpacingRow, isLikelyWideWordSpacingRow, shouldInsertSemanticSpace } from './spacing.js';
 import { isRtlDominantPositionedText, textOrder } from './textDirection.js';
-
-/**
- * Inputs the processor builds once per request, then passes to
- * {@link searchPage} for every page.
- */
-export interface CompiledSearch {
-  /** Each query's matcher, in the same order as the original input.
-   *  Captured as RegExp so literal and regex paths share the iteration
-   *  code below — literal queries get the substring escaped and the
-   *  `g` flag added, regex queries get the user pattern verbatim. */
-  matchers: { query: string; regex: RegExp; queryIndex?: number }[];
-  /** Whether NFKC normalization applies on the *document* side
-   *  (spans / OCR haystack). Literal-mode queries are also NFKC-
-   *  normalised in this case so `"fi"` finds compatibility ligature
-   *  `"ﬁ"` (U+FB01) PDFs. Regex queries are NEVER normalised, even when this flag
-   *  is true — NFKC can turn compatibility punctuation into regex
-   *  metacharacters; users opting into regex get literal codepoints
-   *  against the normalised document text and own the asymmetry.
-   *  When `--no-normalize` is on, this is false and the document
-   *  side stays raw too. */
-  normalize: boolean;
-}
-
-/**
- * Escape every character that has special meaning in a JavaScript
- * RegExp so a literal-mode query like `"3.14"` doesn't accidentally
- * match `"3X14"`. Used only on the literal path; regex-mode queries
- * skip this and are compiled verbatim.
- */
-function escapeRegExp(s: string): string {
-  return s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-}
-
-function literalSearchPattern(s: string): string {
-  const chars = Array.from(s);
-  let pattern = '';
-  for (let i = 0; i < chars.length; i++) {
-    if (i > 0 && isCjkLeading(chars[i - 1]) && isCjkLeading(chars[i])) {
-      pattern += '\\s*';
-    }
-    pattern += escapeRegExp(chars[i]);
-  }
-  return pattern;
-}
-
-/** Apply NFKC the same way `processor.normalizeText` does. Inlined
- *  here so search.ts doesn't reach back into processor.ts. */
-function nfkc(s: string): string {
-  return s.normalize('NFKC');
-}
-
-/**
- * Compile the user-supplied search inputs into a {@link CompiledSearch}
- * the processor can reuse across pages. Throws on:
- *   - empty array / empty string queries (nothing to search for)
- *   - invalid regex when `regex` is true
- *
- * Single-query callers omit `queryIndex` on the resulting matchers so
- * downstream `SearchMatch.queryIndex` stays undefined for the common
- * case; multi-query callers get 0-based indices.
- */
-export function compileSearch(
-  search: string | string[] | undefined,
-  options: { regex?: boolean; caseSensitive?: boolean; normalize?: boolean },
-): CompiledSearch | undefined {
-  if (search === undefined) return undefined;
-  const queries = Array.isArray(search) ? search : [search];
-  if (queries.length === 0) {
-    throw new Error('search: expected at least one query');
-  }
-  for (const q of queries) {
-    if (typeof q !== 'string' || q.length === 0) {
-      throw new Error('search: query must be a non-empty string');
-    }
-  }
-  const normalize = options.normalize !== false;
-  const isMulti = queries.length > 1;
-  const flags = options.caseSensitive ? 'g' : 'gi';
-  const matchers = queries.map((rawQuery, i) => {
-    let pattern: string;
-    if (options.regex) {
-      // Verbatim — let JS RegExp surface invalid-pattern errors with
-      // their own messages. Crucially we do NOT NFKC-normalize regex
-      // queries: NFKC can turn compatibility punctuation into regex
-      // metacharacters (`Ａ．Ｂ` → `A.B`, silent overmatch) or break
-      // syntax outright (`［…］` → `[…]`, may not be a valid char
-      // class). Users opting into regex semantics get the literal
-      // codepoints they typed.
-      pattern = rawQuery;
-    } else {
-      // Literal mode: NFKC the query so `"fi"` finds compatibility
-      // ligature `"ﬁ"` (U+FB01) PDFs, matching what we do to the
-      // document text. Escape *after* normalization so a fullwidth
-      // dot that normalises to `.` is still treated literally.
-      const query = normalize ? nfkc(rawQuery) : rawQuery;
-      pattern = literalSearchPattern(query);
-    }
-    let regex: RegExp;
-    try {
-      regex = new RegExp(pattern, flags);
-    } catch (err) {
-      throw new Error(
-        `Invalid search query ${JSON.stringify(rawQuery)}: ${err instanceof Error ? err.message : String(err)}`,
-      );
-    }
-    return { query: rawQuery, regex, ...(isMulti && { queryIndex: i }) };
-  });
-  return { matchers, normalize };
-}
 
 /** Round to 2dp — matches the convention used by spans / layout / region. */
 function round2(n: number): number {
@@ -156,7 +53,6 @@ const VERTICAL_SEARCH_MIN_SPAN_HEIGHT_RATIO = 2;
 const LATIN_OR_NUMBER_END_RE = /[\p{Script=Latin}\p{M}\p{N}]$/u;
 const LATIN_OR_NUMBER_START_RE = /^[\p{Script=Latin}\p{M}\p{N}]/u;
 const LOWERCASE_START_RE = /^\p{Ll}/u;
-const PRECISE_DUPLICATE_MIN_OVERLAP_RATIO = 0.5;
 
 interface SearchLine {
   text: string;
@@ -174,87 +70,6 @@ interface Box {
 
 interface SearchOwner extends Box {
   text: string;
-}
-
-function duplicateKey(queryIndex: number | undefined, query: string, text: string, ignoreCase: boolean): string {
-  const queryKey = queryIndex === undefined ? query : String(queryIndex);
-  const textKey = ignoreCase ? text.toLowerCase() : text;
-  return `${queryKey}\u0000${textKey}`;
-}
-
-function matcherForMatch(compiled: CompiledSearch, match: SearchMatch): { regex: RegExp } | undefined {
-  if (match.queryIndex !== undefined) {
-    return compiled.matchers.find((m) => m.queryIndex === match.queryIndex && m.query === match.query);
-  }
-  return compiled.matchers.find((m) => m.query === match.query);
-}
-
-function duplicateKeyForMatch(compiled: CompiledSearch, match: SearchMatch): string {
-  const matcher = matcherForMatch(compiled, match);
-  return duplicateKey(match.queryIndex, match.query, match.text, matcher?.regex.ignoreCase ?? false);
-}
-
-function preciseDuplicateBudget(
-  preciseMatches: readonly SearchMatch[] | undefined,
-  compiled: CompiledSearch,
-): Map<string, number> {
-  const budget = new Map<string, number>();
-  for (const match of preciseMatches ?? []) {
-    if (match.source === 'ocr') continue;
-    const key = duplicateKeyForMatch(compiled, match);
-    budget.set(key, (budget.get(key) ?? 0) + 1);
-  }
-  return budget;
-}
-
-function hasPreciseDuplicateAtBox(
-  preciseMatches: readonly SearchMatch[] | undefined,
-  compiled: CompiledSearch,
-  key: string,
-  box: Box,
-): boolean {
-  for (const match of preciseMatches ?? []) {
-    if (match.source === 'ocr') continue;
-    if (duplicateKeyForMatch(compiled, match) !== key) continue;
-    if (boxOverlapRatio(match.bbox, box) >= PRECISE_DUPLICATE_MIN_OVERLAP_RATIO) return true;
-  }
-  return false;
-}
-
-function boxOverlapRatio(a: Box, b: Box): number {
-  const areaA = Math.max(0, a.width) * Math.max(0, a.height);
-  const areaB = Math.max(0, b.width) * Math.max(0, b.height);
-  const smallerArea = Math.min(areaA, areaB);
-  if (smallerArea <= 0) return 0;
-  const x1 = Math.max(a.x, b.x);
-  const y1 = Math.max(a.y, b.y);
-  const x2 = Math.min(a.x + a.width, b.x + b.width);
-  const y2 = Math.min(a.y + a.height, b.y + b.height);
-  const overlap = Math.max(0, x2 - x1) * Math.max(0, y2 - y1);
-  return overlap / smallerArea;
-}
-
-export function suppressDuplicateOcrMatches(
-  nativeMatches: readonly SearchMatch[] | undefined,
-  ocrMatches: readonly SearchMatch[],
-  compiled: CompiledSearch,
-): SearchMatch[] {
-  const budget = preciseDuplicateBudget(nativeMatches, compiled);
-  const out: SearchMatch[] = [];
-  for (const match of ocrMatches) {
-    if (match.source !== 'ocr') {
-      out.push(match);
-      continue;
-    }
-    const key = duplicateKeyForMatch(compiled, match);
-    const remaining = budget.get(key) ?? 0;
-    if (remaining > 0) {
-      budget.set(key, remaining - 1);
-      continue;
-    }
-    out.push(match);
-  }
-  return out;
 }
 
 function buildSearchLines(spans: readonly TextSpan[] | undefined, pageWidth: number): SearchLine[] {
